Extract `fj-host` from `fj-app`
This is the plugin host code for the models. Recycles the `fj-host`
crate name.
