+++ fj-host/Cargo.toml
@@ -4,6 +4,17 @@ version = "0.5.0"
 edition = "2021"
 
 description = "The world needs another CAD program."
-readme      = "README.md"
+readme      = "../README.md"
 repository  = "https://github.com/hannobraun/fornjot"
 license     = "0BSD"
+keywords    = ["cad", "programmatic", "code-cad"]
+
+
+[dependencies]
+libloading = "0.7.2"
+notify     = "5.0.0-pre.14"
+thiserror  = "1.0.30"
+
+[dependencies.fj]
+version = "0.5.0"
+path    = "../fj"
