--- fj-host/README.md
@@ -1,1 +1,0 @@
-`fj-host` has been renamed to `fj-app`. Please refer to the [`fj-app` crate](https://crates.io/crates/fj-app).
